Add: 상수데이터, Mock 데이터 구현
