+++ public/data/userInfoList.json
@@ -1,0 +1,6 @@
+[
+  { "id": 1, "text": "피드1", "img": "/images/img.boyoonkim/boimg.jpg" },
+  { "id": 2, "text": "피드2", "img": "/images/img.boyoonkim/dogimg.jpg" },
+  { "id": 3, "text": "피드3", "img": "/images/img.boyoonkim/dogimg2.jpg" },
+  { "id": 4, "text": "피드4", "img": "/images/img.boyoonkim/smile img.jpg" }
+]
